Flattening RuleResolver
